Extract transaction history logic to own class/file - also changes the way metadata are created so that that address discovery and transaction history are totally separated
