+++ src/crypto/wallet.js
@@ -1,13 +1,10 @@
 // @flow
 
-import moment from 'moment'
 import _ from 'lodash'
 import {BigNumber} from 'bignumber.js'
-import {defaultMemoize} from 'reselect'
 
 import {AddressChain, AddressGenerator} from './chain'
 import * as util from './util'
-import {ObjectValues} from '../utils/flow'
 import api from '../api'
 import {CONFIG} from '../config'
 import assert from '../utils/assert'
@@ -16,81 +13,18 @@ import {
   synchronize,
   nonblockingSynchronize,
   IsLockedError,
-  limitConcurrency,
 } from '../utils/promise'
-import {TRANSACTION_STATUS} from '../types/HistoryTransaction'
+import {TransactionCache} from './transactionCache'
 
-import type {Moment} from 'moment'
 import type {
-  Transaction,
   RawUtxo,
   TransactionInput,
   PreparedTransactionData,
 } from '../types/HistoryTransaction'
-import type {Dict} from '../state'
-
 import type {Mutex} from '../utils/promise'
 
-const getLastTimestamp = (history: Array<Transaction>): ?Moment => {
-  // Note(ppershing): ISO8601 dates can be sorted as strings
-  // and the result is expected
-  return _.max(history.map((tx) => tx.lastUpdatedAt), moment(0))
-}
-
-type SyncMetadata = {
-  lastUpdated: Moment,
-  bestBlockNum: number,
-}
-
 type WalletHistoryState = {
-  transactions: Dict<Transaction>,
-  perAddressSyncMetadata: Dict<SyncMetadata>,
   generatedAddressCount: number,
-}
-
-const perAddressTxsSelector = (state: WalletHistoryState) => {
-  const transactions = state.transactions
-  const addressToTxs = {}
-  const addTxTo = (txId, addr) => {
-    const current = addressToTxs[addr] || []
-    const cleared = current.filter((_txId) => txId !== _txId)
-    addressToTxs[addr] = [...cleared, txId]
-  }
-
-  ObjectValues(transactions).forEach((tx) => {
-    tx.inputs.forEach(({address}) => {
-      addTxTo(tx.id, address)
-    })
-    tx.outputs.forEach(({address}) => {
-      addTxTo(tx.id, address)
-    })
-  })
-  return addressToTxs
-}
-
-const confirmationCountsSelector = (state: WalletHistoryState) => {
-  const {perAddressSyncMetadata, transactions} = state
-  return _.mapValues(transactions, (tx: Transaction) => {
-    if (tx.status !== TRANSACTION_STATUS.SUCCESSFUL) {
-      // TODO(ppershing): do failed transactions have assurance?
-      return null
-    }
-
-    const getBlockNum = ({address}) =>
-      perAddressSyncMetadata[address]
-        ? perAddressSyncMetadata[address].bestBlockNum
-        : 0
-
-    const bestBlockNum = _.max([
-      tx.bestBlockNum,
-      ...tx.inputs.map(getBlockNum),
-      ...tx.outputs.map(getBlockNum),
-    ])
-
-    assert.assert(tx.blockNum, 'Successfull tx should have blockNum')
-    /* :: if (!tx.blockNum) throw 'assert' */
-    return bestBlockNum - tx.blockNum
-  })
 }
 
 export class WalletManager {
@@ -103,9 +37,8 @@ export class WalletManager {
   _isInitialized: boolean
   _doFullSyncMutex: Mutex
   _restoreMutex: Mutex
-  _perAddressTxsSelector: any
-  _confirmationCountsSelector: any
   _subscriptions: Array<() => any>
+  _transactionCache: TransactionCache
 
   constructor() {
     this._encryptedMasterKey = null
@@ -122,10 +55,8 @@ export class WalletManager {
     this._isInitialized = false
     this._doFullSyncMutex = {name: 'doFullSyncMutex', lock: null}
     this._restoreMutex = {name: 'restoreMutex', lock: null}
-    this._perAddressTxsSelector = defaultMemoize(perAddressTxsSelector)
-    this._confirmationCountsSelector = defaultMemoize(
-      confirmationCountsSelector,
-    )
+    this._transactionCache = new TransactionCache()
+    this._transactionCache.subscribe(() => this.updateState({}))
     this._subscriptions = []
   }
 
@@ -146,20 +77,12 @@ export class WalletManager {
     this._subscriptions.push(handler)
   }
 
-  get perAddressTxs() {
-    return this._perAddressTxsSelector(this._state)
-  }
-
   get transactions() {
-    return this._state.transactions
+    return this._transactionCache.transactions
   }
 
   get confirmationCounts() {
-    return this._confirmationCountsSelector(this._state)
-  }
-
-  _getAccount(masterKey) {
-    return util.getAccountFromMasterKey(masterKey)
+    return this._transactionCache.confirmationCounts
   }
 
   restoreWallet(mnemonic: string, newPassword: string) {
@@ -172,7 +95,7 @@ export class WalletManager {
     Logger.info('restore wallet')
     assert.assert(!this._isInitialized, 'restoreWallet: !isInitialized')
     const masterKey = await util.getMasterKeyFromMnemonic(mnemonic)
-    const account = await this._getAccount(masterKey)
+    const account = await util.getAccountFromMasterKey(masterKey)
     this._encryptedMasterKey = await util.encryptMasterKey(
       newPassword,
       masterKey,
@@ -207,7 +130,6 @@ export class WalletManager {
   }
 
   onDiscoveredAddresses(addresses: Array<string>) {
-    this.createMetadataForAddresses(addresses)
     // broadcast change
     this.updateState({})
   }
@@ -253,13 +175,13 @@ export class WalletManager {
     Logger.info('Discovery done, now syncing transactions')
     let keepGoing = true
     while (keepGoing) {
-      keepGoing = await this.doSyncStep([
+      keepGoing = await this._transactionCache.doSyncStep([
         ...this._internalChain.getBlocks(),
         ...this._externalChain.getBlocks(),
       ])
     }
 
-    return {...this._state.transactions}
+    return this._transactionCache.transactions
   }
 
   // TODO(ppershing): memoize
@@ -293,7 +215,8 @@ export class WalletManager {
 
   isUsedAddress(address: string) {
     return (
-      !!this.perAddressTxs[address] && this.perAddressTxs[address].length > 0
+      !!this._transactionCache.perAddressTxs[address] &&
+      this._transactionCache.perAddressTxs[address].length > 0
     )
   }
 
@@ -314,146 +237,6 @@ export class WalletManager {
     })
 
     return true
-  }
-
-  _isUpdatedTransaction(tx: Transaction): boolean {
-    const id = tx.id
-    // We have this transaction and it did not change
-    if (
-      this._state.transactions[id] &&
-      this._state.transactions[id].lastUpdatedAt.isSame(tx.lastUpdatedAt)
-    ) {
-      return false
-    }
-    if (this._state.transactions[id]) {
-      // Do things that matter if the transaction changed!
-      Logger.info('Tx changed', tx)
-    }
-    return true
-  }
-
-  // Returns number of updated transactions
-  _checkUpdatedTransactions(transactions: Array<Transaction>): number {
-    Logger.debug('_updateTransactions', transactions)
-    // Currently we do not support two updates inside a same batch
-    // (and backend shouldn't support it either)
-    assert.assert(
-      transactions.length === _.uniq(transactions.map((tx) => tx.id)).length,
-      'Got the same transaction twice in one batch',
-    )
-    const updated = transactions.map((tx) => this._isUpdatedTransaction(tx))
-    return _.sum(updated, (x) => (x ? 1 : 0))
-  }
-
-  createMetadataForAddresses(addrs: Array<string>) {
-    addrs.forEach((addr) => {
-      assert.assert(
-        _.isUndefined(this._state.perAddressSyncMetadata[addr]),
-        'createMetadataForAddresses: new addresses',
-      )
-    })
-
-    const newMetadata = _.fromPairs(
-      addrs.map((addr) => [
-        addr,
-        {
-          lastUpdated: moment(0),
-          bestBlockNum: 0,
-        },
-      ]),
-    )
-
-    this.updateState({
-      perAddressSyncMetadata: {
-        ...this._state.perAddressSyncMetadata,
-        ...newMetadata,
-      },
-    })
-  }
-
-  _getBlockMetadata(addrs: Array<string>) {
-    assert.assert(addrs.length, 'getBlockMetadata: addrs not empty')
-    const metadata = addrs.map(
-      (addr) => this._state.perAddressSyncMetadata[addr],
-    )
-    // check consistency
-    assert.assert(
-      metadata,
-      'getBlockMetadata: metadata should not be undefined',
-    )
-    assert.assert(
-      metadata.every((x) => x.lastUpdated.isSame(metadata[0].lastUpdated)),
-      'getBlockMetadata: lastUpdated metadata same',
-    )
-    assert.assert(
-      metadata.every((x) => x.bestBlockNum === metadata[0].bestBlockNum),
-      'getBlockMetadata: bestBlockNum metadata same',
-    )
-
-    return metadata[0]
-  }
-
-  async doSyncStep(blocks: Array<Array<string>>): Promise<boolean> {
-    Logger.info('doSyncStep', blocks)
-    let count = 0
-    let wasPaginated = false
-    const errors = []
-
-    const tasks = blocks.map((addrs) => {
-      const metadata = this._getBlockMetadata(addrs)
-      return () =>
-        api
-          .fetchNewTxHistory(metadata.lastUpdated, addrs)
-          .then((response) => [addrs, response])
-    })
-
-    const limit = limitConcurrency(CONFIG.MAX_CONCURRENT_REQUESTS)
-
-    const promises = tasks.map((t) => limit(t))
-
-    // Note(ppershing): This serializes the respons order
-    // but still allows for concurrent requests
-    for (const promise of promises) {
-      try {
-        const [addrs, response] = await promise
-        wasPaginated = wasPaginated || !response.isLast
-        const metadata = this._getBlockMetadata(addrs)
-        const newLastUpdated = getLastTimestamp(response.transactions)
-        // Note: we can update best block number only if we are processing
-        // the last page of the history request, see design doc for details
-        const newBestBlockNum =
-          response.isLast && response.transactions.length
-            ? response.transactions[0].bestBlockNum
-            : metadata.bestBlockNum
-
-        const newMetadata = {
-          lastUpdated: newLastUpdated,
-          bestBlockNum: newBestBlockNum,
-        }
-
-        const transactionsUpdate = _.fromPairs(
-          response.transactions.map((tx) => [tx.id, tx]),
-        )
-        const metadataUpdate = _.fromPairs(
-          addrs.map((addr) => [addr, newMetadata]),
-        )
-
-        count += this._checkUpdatedTransactions(response.transactions)
-
-        this.updateState({
-          transactions: {...this._state.transactions, ...transactionsUpdate},
-          perAddressSyncMetadata: {
-            ...this._state.perAddressSyncMetadata,
-            ...metadataUpdate,
-          },
-        })
-      } catch (e) {
-        errors.push(e)
-      }
-    }
-
-    if (errors.length) throw errors
-    return wasPaginated || count > 0
   }
 
   transformUtxoToInput(utxo: RawUtxo): TransactionInput {
